refactor: modernize codebase with C++20/23 features and improve error handling
- Replace raw arrays with std::array for chunk blocks and frame tracking.
- Convert VoxelMesh vertices and indices from static std::vector to header-only static constexpr std::array and remove VoxelMesh.cpp.
- Update FileSystem to use std::expected for robust error handling.
- Modernize ThreadPool to use std::move_only_function for tasks.
- Adopt modern C++ ranges and algorithms (std::ranges::fold_left, std::ranges::all_of).
- Optimize enum sizes for InputAction, MouseCode, and KeyCode.
- Use std::scoped_lock instead of lock_guard in World and MegaBuffer.
- Clean up raw pointer initialization and C-style casts.
- Add CMake hints for finding clang-tidy in Visual Studio LLVM directories.

+++ src/input/InputAction.h
@@ -4,7 +4,7 @@
 
 namespace voxl
 {
-    enum class InputAction : uint32_t
+    enum class InputAction : uint8_t
     {
         MoveForward,
         MoveBackward,
